This is a new example for "STITLE", illustrating the new features that I added to the code.
This example also illustrates how to use FTITLE, which the previous
examples didn't.  For simplicity, instead of trying to build the logic
for providing an FTITLE input file into the distribution scheme, I built
into the example itself a little subroutine that writes the input file
FTITLE will be told to read.

This example also illustrates the use of a user version of the routine
SLUBKG, which can be made to put a background behind a scrolled title.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3553 10fe2890-8a40-4696-8b1d-7ccbb19c398e