Miscellanous bug fixes.
ContourPlot: removed a line that prevented ContourPlot from responding
to changes to the tr[XY][Min/Max]F variables when there is no contour data.

LabelBar/Legend: fixed a problem that allowed the geometry state
variables to get screwed up if the viewport was modified while the
object was turned off.

MapPlot: the wild card character ('*') was not be interpreted correctly
when it occurred at the end of a string in a specification list.

PlotManager: fixed a problem with updating the TickMark viewport when
an overlay is removed from the base plot (or the base plot is destroyed).
Also added some code that restores the viewport of an overlay when
it is removed from the base or the base is destroyed. The viewport is
returned to the state it had when the overlay was added.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3555 10fe2890-8a40-4696-8b1d-7ccbb19c398e