Used the Areas 'RC' parameter to fix the problem Mary found with map fill for LambertConformal projections.
Also in MapPlot: beefed up error checking for line color index; now defaults
to Foreground if negative (Transparent is not allowed for lines drawn
by Ezmap).

In MapTransObj: I have modified the transformation functions to properly
handle longitudes in the range -540 to +540. I overloaded the mpMinLonF
and mpMaxLonF resources to allow you to specify the range of the data
you are expecting. This makes it possible to overlay a ContourPlot that
has data in the range from 0 to 360 (for instance). Possibly I still might
decide that there should be new resources to specify the expected data
range.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3593 10fe2890-8a40-4696-8b1d-7ccbb19c398e