Added a new parameter called VRM (Vector reference magnitude). This parameter allows the user to specify a reference magnitude that is smaller than the maximum displayed magnitude. Formerly, this was not possible. The reference magnitude may be any positive value, even less than the minimum vector magnitude. When VRM is 0.0, its default value, Vectors works just as it did previously.
I modified all the SETER calls to make them recoverable, and replaced
all STOPs with RETURNs.

I also fixed a couple of bugs: a problem with the vector arrow sometimes
overlapping the text in the vector reference length annotations; and
also a problem with extra spaces being included in the vector debugging
labels.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3633 10fe2890-8a40-4696-8b1d-7ccbb19c398e