This will fix a problem reported by Tom Parker and Mary Haley.
In CPTRCL, CPTREG, and CPTRES, I changed some code that is intended to
make a contour line that passes very close to a grid point be treated as
if it passed through the grid point.  This prevents the creation in the
contour line of tiny segments that cause problems for other routines; it
was being done in such a way as to create problems when using a very large
grid.

In CPPKCL, CPSP1A, and CPSP2A, I just made some minor changes in the way
a few constants were being written.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3644 10fe2890-8a40-4696-8b1d-7ccbb19c398e