Initial version of VectorPlot and VectorField public classes. Also of the VecAnno class used privately by the VectorPlot class.
Also some miscellaneous fixes and changes.

ContourPlot: fixed a bug that caused an infinite loop under certain
conditions when the scaling mode was set to INTEGERLINELABELS. Also
changed the behavior of the LabelScaleFactorF resource to match the
documentation. It had been acting as a divisor rather than as a
multiplier. I made the types NhlcnLevelSelectionMode and NhlcnLabelScalingMode
into synonyms for the new more general DataComm types NhlLevelSelectionMode
and NhlScalingMode.

DataComm: Added new types NhlLevelSelectionMode and NhlScalingMode. These
are used both by VectorPlot and ContourPlot.

ScalarField: Added new read-only resources sf{XY}CActual{Start|End}. These
give the actual data coordinate start and end values when subsetting
or strides are in effect.
Also modified the effect of setting the sfExchangeDimensions resource True.
The sfXC... and sfYC.. resources now refer to the original dimensions of
the array (sfXC... resources always refer to the fast varying dimension
of the original array). This means that coordinate extents defined using
sfXC... resources will appear along the Y-Axis of a ContourPlot when
ExchangeDimensions is set True.
Also fixed a bug involving the data min and data max values.

TextItem,Transform, and View: replaced some instances of the obsolete
preprocessor define "NhlNeedFuncProto" with the current symbol "NhlNeedProto".


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3653 10fe2890-8a40-4696-8b1d-7ccbb19c398e