Hi John,
I took a look at this since I would like to modify the way colors are allocated
in ctrans and idt.  I'm pretty sure I have figured out what was going on
here, but could you take a look at it and let me know if you agree?

-----------------------
Fixed a bug in SetInPic.  If the default color table was currently
"damaged", SetInPic wasn't marking the new "working" color table "damaged".
So, X11_UpdateColorTable_ wasn't being called even though there were
"damaged" elements in the color table.

This was showing itself in that the -pal option to ctrans wasn't working.
As far as I can tell this is the only time that ColrTable gets called
modifying the default color table.  Usually it is called to modify the
"working" color table.  ColrTable is the function that marks the
Color Table as "damaged" allowing X11_UpdateColorTable_ to be called
later, so as long as it was called on the "working" table and not the
default one between pictures, everything would work ok.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3668 10fe2890-8a40-4696-8b1d-7ccbb19c398e