This checkin adds new options to VectorPlot: variable width filled arrows and the ability to thin vectors based on a minimum NDC distance. It requires newly checked in LLU Vectors code to work.
It also fixes a problem discovered by Ben Foster: exponential formatting
did not work properly when the TextItem function code was changed. The
fix involved adding a function code parameter to _NhlFormatFloat. Also
I added new function code resources in TickMark and to the private
MultiText class. In addition I had to intercept the LabelBar and Legend
resources (lb/lg)LabelFuncCode in ContourPlot and VectorPlot.

ContourPlot: modified all calls to _NhlFormatFloat; intercepted
(lb/lg)LabelFuncCode.

Format: modified _NhlFormatFloat to allow the caller to specify the func
code.

MultiText: added a FuncCode resource


PlotManager: modified TickMark code so that the transformation is not
modified unecessarily when Ticks are turned off because a plot is added to
an overlay.

TickMark: added four FuncCode resources (one for each axis) and modified
call chain to support the new function code parameter in _NhlFormatFloat.

VecAnno: modified to support private communication with LLU Vectors code.
This is necessary in order to ensure that Vectors is in the proper state
when it sizes and draws the annotation arrow.

VectorPlot: added functionality described above; modified calls to
_NhlFormatFlost.

Workspace: added wrapper functions _NhlVvinit/_NhlVvectr and
_NhlStinit/_NhlStream to support Vectors and Streamlines workspaces.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3730 10fe2890-8a40-4696-8b1d-7ccbb19c398e