New functionality:
Modified NCL such that variables can be read out of file with out nowing
the names of the variables when writing the script. A new function called
"getfilevarnames" is needed to use the new functionality. getfilevarnames
returns all of the variable names for a file. Now by using ->$STRING$ instead
of ->var_name you can reference a variable in a file by a string rather
than hard codeing it. For Example:

The following script print all of the variable in a file.

	a = addfile("testit.cdf","r")
	vn = getfilevarnames(a)
	do i = 0,dimsizes(vn)-1
		print(a->$vn(i)$)
	end do


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3745 10fe2890-8a40-4696-8b1d-7ccbb19c398e