This checkin implements immediate mode drawing methods, along with a new Style class for maintaining the attributes of drawing primitives. The interface for the existing NDC/Data Polyline methods is modified to include a GraphicStyle object as a parameter.
NDC/Data Polygon and Polymarker methods have been added to the Transform
class -- requiring modifications to the class initializer of each of the
Tranform subclasses.

The Style class currently contains 5 subclasses - the subclasses for
the individual primitives: LineStyle, MarkerStyle, FillStyle, and TextStyle,
as well as the GraphicStyle subclass, a union of each of the individual
primitive subclasses.

When a Workstation is initialized it creates a default GraphicStyle: this
style is used if the immediate mode primitives are called with object id
0 (NhlNULLOBJID, or a define still to be added - NhlDEFAULTSTYLE).

The "wk" line and marker primitive resources are now considered obsolete.
If the user sets them, a warning will be issued, but, for now, they
will be used to set corresponding resources in the default graphic
style object. At 4.1, it is anticipated that these resources will be
removed entirely.

A new class flag has been added for Style in hluP.h -- so a complete
re-compile of the HLU library is required -- sorry.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3767 10fe2890-8a40-4696-8b1d-7ccbb19c398e