This update is in response to a perennial complaint about CONPACK - one heard often from the consultants and again at the user conference.
The changes make it possible for a user to provide his/her own versions
of the routines CPCHHL and CPCHLL with code to suppress the placement of
labels in areas in which labels are not desired (as, for example, over
land on an EZMAP background, if the field being contoured is intrinsically
defined only over water).

As a secondary effect of these changes, the routines CPHLLB and CPINLB
have been considerably simplified.  Previously, they had code in them
allowing the high/low labels and the informational label to be drawn
or to be put into an area map directly.  This code was executed when
contour line labels were being generated by the dash package (that is,
when 'LLP' had the value 1 instead of 2 or 3) and saved a little space
in the real workspace array.  Now, no matter what the value of 'LLP',
descriptors of the high/low labels and the informational label will
be generated and put in a list in the real workspace array for later
use by the routines CPLBAM and CPLBDR.  Although the new code requires
a little more user space, it will be easier to check out and to maintain,
as it results in less duplicated code in different CONPACK routines.

I am about to check in an example illustrating how the routines CPCHHL
and CPCHLL may now be used.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3774 10fe2890-8a40-4696-8b1d-7ccbb19c398e