Streamlines now displays a ZERO FIELD message when presented with a zero field. Error recovery implemented.
I added several new parameters: 'GBS' (grid-based spacing) is a flag that
allows the user to modify the interpretation of several parameters.
'AMD' allows the user to specify a minimum distance separating the
directional arrows on a single streamline. There are also parameters to
control the ZERO FIELD message: these are exactly the same as in Vectors.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3802 10fe2890-8a40-4696-8b1d-7ccbb19c398e