Mary, this fixes trn00023 (ctrans bombing on TrueColor display's).
I have added an undocumented feature to ctrans.  The "-visual id" option is now
available for use with the X11 driver.  This is used to specify the X
visual you want ctrans to use to create the Xwindow.  This is not intended to
be a user option - I put it in mostly so I could easily test that ctrans
would work with a given visual.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3816 10fe2890-8a40-4696-8b1d-7ccbb19c398e