Fixed bugs 49 and 51.
Coordinate subscripting is fixed.

NCL still doesn't allow strings to be written to netCDF but it does
now provide a resonable message and suggest that the appropriate
type conversion function be used. I add stringtochar and chartostring
so that variables can be converted back and forth manually by the user.

	-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3888 10fe2890-8a40-4696-8b1d-7ccbb19c398e