Major bug fixes.
I now use Jeff's callback interface to keep track of HLU objects in NCL.
This was a major problem with 4.0. NCL never new when to destroy an HLU
object and when not to. Now callbacks make sure that all MultDValHLUObjData's
are notified when HLU objects they reference are destroyed. This happens when
NhlDestroy is called or when a parent of and HLU object is destroyed
(ie workstations or app objects). Also now NCL knows not to destroy objects
(unless NhlDestroy) is called that have been assigned to a resource either
through the setvalues, create, NhlAddAnnotation, or NhlAddData calls. These
types of objects are now considered "exported" objects and the objects
they've whose resources they've been assigned to maintain a reference to
these objects, preventing their destruction. When the objects that reference
these are destroy so are the exported object, provided no other variable
references these objects.

This is a good solution but it does allow in some instances HLU objects that
are not being used to remain instantiated longer than they should.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3930 10fe2890-8a40-4696-8b1d-7ccbb19c398e