Fixed bug that Ethan found that caused unnecessary draws when in segment mode under some conditions when a ContourPlot was an overlay. Fixed similar bugs in VectorPLot and StreamlinePlot.
Removed the unnecessary restriction that vpUseSegments could only be
set at create in all the View classes that support segments.

Changed the reference view size for ContourPlot, MapPlot, and PlotManager
to be equal to the default view size (instead of 0.5,0.5 as previously
had been the case). Modified the default values of a number of resources
that are adjusted dynamically based on the reference view (if not set
explicitly). The result is that the documented default value should be
be the NDC size you get for these resources if the plot is drawn at the
default size. Most values were adjusted so that the no change will be
noticed in existing examples. However, a few of these changes affect
the appearance of several existing examples.
The most noticeable is probably the change in the default dash segment
length for MapPlot. This is apparent in the mp05 example. It should
probably be modified to compensate.

Also bumped up the value of the Workspace resource wsThresholdSize from
1 meg to 4 meg. Hopefully this will improve performance a bit on most
systems.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3940 10fe2890-8a40-4696-8b1d-7ccbb19c398e