Updates to drawing from segments to work properly in the context of overlays. Now, in segment mode a list of resources is checked at each set value call. The list contains all the resources that can change without needing a complete redraw. If the arg list contains any resources not in this list, a redraw will take place. Otherwise the plot will be drawn from the segment. Some code was added to PlotManager that helps avoid unnecessary redraws when overlays are added or removed.
Unfortunately, there is a problem with VectorPlot segments. VectorPlot
turns clipping on while drawing so that all the arrows are clipped to
the viewport. However, apparently the clipping rectangle does not
transform along with the rest of the segment, so that if you move a
VectorPlot segment, the plot becomes partially or fully invisible.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3973 10fe2890-8a40-4696-8b1d-7ccbb19c398e