This fixes a problem Dave found with the missing values.  The my_missing_? field wasn't being reset when the corresponding missing_? value was changed.
Dave, your solution introduced a memory leak.  FlushObj can be called many
times between changes in the missing value.  If the missing value was set
to a non-float value, then the conversion to float would happen each time
FlushObj is called, and the new my_missing_? gen array would over-write
the existing one causing a memory leak, so the FlushObj function should
only do the conversion if the my_missing_? is NULL, and the missing_?
is set.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4011 10fe2890-8a40-4696-8b1d-7ccbb19c398e