Implemented transformation of polylines and polygon edges for log axes in the LogLinTransObj.
Added a new resource to TransObj that specifies the sampling resolution
as an NDC value: trResolutionF. Setting the resource to 1.0 or greater
ensures that no sampling along the length of the line occurs, and
therefore only the endpoints are transformed. Note that currently the
sampling points are equally spaced in data space (not in NDC). Therefore,
the NDC value is only the approximate average spacing in NDC for each
segment of a polyline or each edge of a polygon.

So far trResolutionF has been implemented in LogLinTransObj and
IrregularTransObj -- still need to do MapTransObj.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4015 10fe2890-8a40-4696-8b1d-7ccbb19c398e