More segment related fixes:
Put _NhlDeleteViewSegments calls into the destroy methods of each of the
View classes listed below, in order to remove segments and segment data
that is around when the view is destroyed.

Fixed a problem in MapTransObj that was causing unnecessary complete draws
when a segment draw should have been sufficient.

Modified the Segment code so that it keeps track of which segment ids are
in use, so that segment ids can be reused. GKS only allows segment ids
1-99, and with *vpUseSegments set true in a resource file, some existing
examples would be getting into the invalid id range.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4055 10fe2890-8a40-4696-8b1d-7ccbb19c398e