Added a Transform class resource 'tfDoNDCOverlay' that causes plot overlays to use their own transobj instead of the base plot's transobj. This allows plots whose data are pre-gridded in a map projection to be overlaid on a map set to the same projection and map limits.
Replaced the Transform interface function _NhlIsOverlayMember with a family
of functions that adhere more precisely to the current interface
terminology. These are:

_NhlIsOverlay -- true if object belongs to Transform class and is currently
 an overlay

_NhlIsAnnotation -- true if object belongs to View class ans is currently an
annotation.

_NhlIsPlotMember -- true if is overlay or is annotation

_NhlOverlayBase -- returns the base plot id of an overlay; its own id if
it is itself an overlay base; or NULLOBJID if not a transform with an
enabled plot manager.

_NhlAnnotationBase -- returns the base plot to which an annotation directly
belongs (not the primary base plot if the base plot is itself an annotation).

_NhlBasePlot -- returns the primary base plot (the drawable object) of
an annotation, an overlay, or the base plot itself.

Also fixed a minor problem in the ncgm workstation.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4178 10fe2890-8a40-4696-8b1d-7ccbb19c398e