Implements the SetValues callback -- "CBresValueSet".
A couple of interface changes were required:
(1) the callback list create function adds two new parameters:
extern _NhlCBList
_NhlCBCreate(
#if     NhlNeedProto
        int             hash_mult,
        _NhlCBAddHash   add_hash,
        _NhlCBCallHash  call_hash,
        _NhlCBTaskProc  task_proc,
        NhlPointer      task_proc_data
#endif
);

task_proc specifies a function that can be used to do private stuff
required for a callback during a callback add, delete, or traversal of a
callback list.
task_proc_data is private data for the task_proc.
If both of these are NULL the callback behavior should remain unchanged.

(2) A field for specifying the task_proc was added to the raw and cooked
object CB records.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4200 10fe2890-8a40-4696-8b1d-7ccbb19c398e