I made a number of changes to fix problems with multiple simulateneous NcgmWorkstations. A temporary close is performed unconditionally on the NcgmWks after a ClearWorkstation call. It is reopened on the next ActivateWorkstation for that wkid. The NcgmWorkstation now has its own versions of all the Workstation methods except for the primitive drawing methods.
I should note that these changes will most likely cause the ncgm output file
sizes for all the ni examples to changes. In all the cases I have observed the
file sizes have been smaller. From looking at the differences in the clear
text output it is not obvious to me why this should be. The graphical output
in the cases I have looked at so far appears the same, but it would probably
be good to compare a little more carefully than usual the example output
with that produced by previous versions of the code.

I replaced calls to c_sflush with c_plotif in Workstation and TickMark. I'm
not sure this was really necessary, given subsequent changes I made, but I
don't think it will hurt.

I also added a slight optimization to the WorkstationAllocateColors routine.
I added a flag that signals if there have been any changes to the color
table since the last call to the routine -- if not the loop through all the
colortable elements is bypassed.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4212 10fe2890-8a40-4696-8b1d-7ccbb19c398e