Renamed varinfogrid* to diminfogrid* to improve naming consistency.
Added call to NgNclGetSymName in order to ensure uniqueness for variables
representing files that are added in the addfile dialog.

Changed NgNclGetSymName so that it returns the input string unmodified
if it represents a unique Ncl symbol name as it stands. (Jeff, I thought you
said that is what this routine was supposed to do. If not, feel free to
back out my change.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4348 10fe2890-8a40-4696-8b1d-7ccbb19c398e