Fixed bug # trn00028:
Description: If you compile and run the following GKS program on the
Crays, ctrans does not display the gmeta file correctly.  The even
fonts seem to be getting reset to odd fonts (for example, font "-4" is
displayed as font "-5"). To compile this program on a Cray, make sure
NCARG_ROOT is set to /usr/local, and type "ncargf77 program.f".


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4460 10fe2890-8a40-4696-8b1d-7ccbb19c398e