These are the routines of the new 3D package called TDPACK.
I have left the following line in the yMakefile:

  PartialLibraryTarget(libNGmisc.a,$(OBJECTS)) ???

I assume this line is intended to add the TDPACK object code to a
particular shared library, but I do not know where it ought to go.
I think my preference would be to put it in a library called

  libNGtd.a

but I have no idea what the pros and cons of doing this might be.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4556 10fe2890-8a40-4696-8b1d-7ccbb19c398e