Implemented a discrete raster contouring option and a mechanism for dynamically adjusting the number of cells in a raster contour based on the resolution of the device. This awaits the implementation of escape functions in the GKS drivers to actually work.
Also modified the level-setting mechanism in ContourPlot and fixed some things
in the float formatting routine.

ContourPlot: added 2 new resources: cnRasterSmoothingOn and
cnRasterSampleFactorF. If cnRasterSmoothingOn is False (currently the default)
then discrete raster contouring will be performed. This means the whole
cell area centered on each grid point will be colored in a single color and
no interpolation to intermediate levels will take place.
cnRasterSampleFactorF provides an alternate mechanism to cnRasterCellSizeF
for setting the number of cells in a raster contour. If set,
cnRasterCellSizeF takes precedence, but it may be unset by giving it the value
0.0. cnRasterSampleFactorF adjusts the number of cells relative to the
default number of cells which is based on the data grid size if
cnRasterSmoothingOn is off and on the device resolution if smoothing is on.

Added the routine _NhlRasterFill to perform non-interpolated raster
contouring.

Level selection was modified to eliminate the switch to AutomaticLevels
mode if the data is out of range of the min and max levels in explicit or
manual mode. Now a warning is simply issued and no contour lines will
appear. If fill is on, the color for above max or below min should fill the
whole plot. Unfortunately this does not work right now for area fill,
because AREAS still has problems with plots that have only a single
undifferentiated area. I will try to develop a work around for this problem.
The motivation for changing things in this way was to make ContourPlot
deal more gracefully with sequences of plots where the data range changes
from plot to plot - for instance, a sequence of temperature plots at
various atmospheric levels. I also tried to eliminate some other instances
of permananent switches to automatic mode. If the setting for Manual or
Explicit levels is invalid, then automatic mode is used to provide some
valid levels, but the mode stays set as it was.

Format: I am checking in changes that allow the user to specify a fixed
number of fraction digits using the ';' and '.' specifiers together. This
also fixes a rounding problem.

Workspace: changed _NhlCpcica to call an hlu routine instead of cpcica when
non-interpolated raster contouring is to be performed.

Workstation: added a get-only resource wkVSWidthDevUnits for retrieving
the the Viewspace width in device units.

XWorkstation: added a GetValues routine for getting the width of the
X viewspace in pixels. This depends on a GKS escape function being
implemented. Right now it simply returns the constant 500.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4603 10fe2890-8a40-4696-8b1d-7ccbb19c398e