Implemented _NhlGetUserResources routine to get a list of all user-accessible resources in each class. This required a new bit field (defined as _NhlRES_PRIVATE) to cull out private resources that should not be visible at the user level. I marked the resources I think should be private in the resource lists of all the publicly-visible classes.
In general the list of resources returned should be those that do not cause
the "xx is not a valid resource in xx at this time" error. However, there
does not seem to be a way to suppress resources for the intrinsic PlotManager
annotations like LabelBar and Legend, even if a plot class
(i.e. StreamlinePlot) unconditionally forbids their use by hard-coding their
display mode to NOCREATE. Maybe StreamlinePlot should not do this. Or maybe
it's up to the GUI level to remove these resources if the user should not
be accessing them.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4616 10fe2890-8a40-4696-8b1d-7ccbb19c398e