Changed the Enum strings to mixed case so they will appear in the GUI in the same 'canonical' form as in the HLU ref docs. Also made similar changes to the string names of the HLU defined types.
Removed local versions of the global _NhlResUnset function.
Replaced individual #defines for the default view width and height with
the View class #defines NHL_DEFAULT_VIEW_WIDTH and NHL_DEFAULT_VIEW_HEIGHT.

Modified the EnumToString converter so that it works for 'hybrid' enums
like ColorIndex that have some named values along with an unnamed range
of integers.

Changed _NhlGetResDB in App.c so that it returns the resource database of
the current app object if a NULL layer is passed in.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4641 10fe2890-8a40-4696-8b1d-7ccbb19c398e