Changes for Solaris/CDE port + other things.
Addfile list translations are now hard-coded because CDE interfered with
resource file translations. I took this opportunity to make a number of
other changes to addfile. The varlist does not change its width nearly
as often, so fewer annoying resizes take place. I moved the label for
the file reference variable name to its own line, and made it (hopefully)
more informative.

DataItems are now accessible from the HLU create menu.

Added special handling in the restree for certain resources that affect
the accessibility of a whole child class. (e.g. pmLabelBarDisplayMode).
This was necessary to avoid the possibility of serious error conditions
occuring.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4736 10fe2890-8a40-4696-8b1d-7ccbb19c398e