This checkin includes interfaces to the XmHTML widget library. You will need paths to the include files and the library to compile libngo and link ngi.
Added support for viewing individual resource documentation using the
XmHTML widget from within the restree.

Added some keyboard traversal to the various tree widgets, along with
auto-scrolling so that as many newly opened rows are in view as is
possible.

Fixed keyboard editing bug in restree, and some sizing and mapping bugs
in the addfile dialogue. There are still some problems in addfile with
user-generated resizes.

Added code so that an update of an overlay plot causes its base plot to
draw rather than generating an error.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4793 10fe2890-8a40-4696-8b1d-7ccbb19c398e