This checkin modifies the yMakefile to require access to the XmHTML widget library as well as the Xpm library. Sgi and Linux seem to find Xpm automatically, but on Solaris it needs to be installed somewhere.
The XmHTML library is used to access the NCARG reference docs. It
can do this either using http and the ngurl value, or it can directly
access the files if available within the filesystem. For the http
method to work an executable version of 'wget' must exist in the user's
path. The filesystem method will take precedence if the
environment variable NCARG_DOC_ROOT is set to the location of the
NCARG documentation ref directory (including the '/' and the path
preceding but not including 'ref'). This method will work if net access
and/or the wget binary is not available; it is also usually faster.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4794 10fe2890-8a40-4696-8b1d-7ccbb19c398e