Implementation of a wind barb mode for Vectors, enabled by setting the AST (Arrow Style) parameter to 2. New parameters are:
WBA - wind barb angle (of the ticks clockwise from the vector direction)
    - make it negative to produce Southern hemisphere wind barbs
    - default: 62.0
WBT - wind barb tick size (as a fraction of the wind barb length)
    - default: 0.33
WBC - wind barb calm circle size (diameter as fraction of wind barb length)
    - default: 0.25
WBD - windbarb distance (spacing between ticks as fraction of wbl -- half
      this distance is used as spacing between pennants.
    - default: 0.1
WBS - windbarb scale factor (scales the vector magnitude -- only for the
      purpose of determining the number of flags and pennants -- does not
      change the reported magnitude) This behavior is still subject to
      possible modification.

Wind barbs are different from the other styles in that they are forced to
a uniform length -- measured as the length of the shaft plus the projection
onto the shaft axis of the top wind barb tick. Also something is drawn
for zero magnitude grid points -- the calm circle. Also note that the
barbs are a composite of polylines and fill areas: the fill area color
is forced to be the same as whatever the line color turns out to be.
Masking to an areamap has not been implemented for wind barbs yet.
In other respects wind barbs share the capabilities of the other styles
including coloring by magnitude and scalar field, setting the reference
length, appearance in the legend, etc.

I also modified the other draw routines (fill and line arrows) a bit in
the interest of efficiency.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4811 10fe2890-8a40-4696-8b1d-7ccbb19c398e