Fixed justification code so that TextItems are justified properly even when they contain position-modifying function codes (C, H, or V). The centering option passed to plotchar is now always set to -1.0. This actually results in the code being quite a bit simpler; it was possible to remove the code branching for the text directions and move a number of statements outside the justification mode switch statement.
I also took the liberty of changing the default text position to 0.5,0.5.
This does not affect any of the example codes, as far as I can tell. It
would allow an alternative version of tx01 that truly sets no non-default
resource values. As it stands, tx01 is forced to set the viewport
explicitly in order to make the whole TextItem appear within the viewspace.
I don't see why the default TextItem should not appear centered in the
viewspace automatically. If someone disagrees, let me know. It can be
changed back to its former state very easily.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4829 10fe2890-8a40-4696-8b1d-7ccbb19c398e