All View objects can now be positioned partially outside the viewspace. They are clipped to the edge of the viewspace. Errors still occur if the View is completely outside the viewspace. I intend to fix this as well. Theoretically at least, vpWidthF and vpHeightF can be set to any positive number. I have tested this with small positive values up to about 4.0. Likewise, vpXF and vpYF can be any real number such that, given the height and width, a portion of the viewport appears inside the viewspace.
When segments are created for a partially visible View object, only the
visible part is drawn into the segment. Therefore, there is new code to
cause a redraw if a view stored as a segment is moved or resized such that
the segment does not contain enough information to create the currently
visible portion of the view.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4880 10fe2890-8a40-4696-8b1d-7ccbb19c398e