Now it only causes an INFO level warning if a view position is set such that its extent is completely outside the viewspace. Calling the view's draw method in such a case is essentially a no-op.
Fixed the behavior of the vpOn resource: if it is set to False for an
overlay member plot, sssts annotations no longer draw when the base plot
is drawn. The base plot repositions its remaining annotations.

Also modified all View classes that implement segments: setting vpOn does
not require a full redraw if a valid segment exists.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4884 10fe2890-8a40-4696-8b1d-7ccbb19c398e