TextItem: added support for segments. ContourPlot: InfoLabel now drawn as a segment if segments are enabled for the contourplot.
A default plot is now created if you pass a workstation id rather than a
plot id as the first parameter of any of the primitive drawing routines.
The plot created is a LogLinPlot with all resources defaulted except for
tfPlotManagerOn, which is set to False. The plot id is private to the
Workstation, preventing the user from modifying this plot's resources or
destroying it. Therefore


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4889 10fe2890-8a40-4696-8b1d-7ccbb19c398e