TextItem: added support for segments. ContourPlot: InfoLabel now drawn as a segment if segments are enabled for the contourplot.
A default plot is now created if you pass a workstation id rather than a
plot id as the first parameter of any of the primitive drawing routines.
The plot created is a LogLinPlot with the viewport set to cover the
viewspace, tfPlotManagerOn set to False, and all other resources defaulted.
The plot id is private to the Workstation, preventing the user from
modifying this plot's resources or destroying it. Drawing to the default
plot is somewhat analogous to using the default normalization transformation
in GKS. There is an identity transformation from Data to NDC coordinates,
so there is no difference between using the NDC and Data versions of the
primitive functions.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4890 10fe2890-8a40-4696-8b1d-7ccbb19c398e