This version of AREDAM clips the incoming edges against the rectangle representing the plotter frame before putting them into the area map.
This is in response to a problem that Dave Brown reported; he was trying
to create an area map containing lines from the new EZMAP database, but
with most of the lines mapping to positions outside the plotter frame; he
found that, although it could be made to work, the size of the area map
required was very large.  When I wrote AREAS, I assumed that such boundary
lines would be clipped against the edges of the plotter frame prior to
calling AREDAM and I therefore put into that routine only some code aimed
at handling minor errors in the clipping; in a case like Dave's, this code
undesirable side effects.

This change is a little scary because it's going in at a pretty low level
and could affect a lot of stuff (in CONPACK, for example).  I have tested
it as thoroughly as I can and it seems to work fine (except for slowing
AREDAM down slightly), but if anyone notices anything funny happing to
area fills that depend on AREAS, please let me know.

The changes in ARGTAI, ARPRAM, and ARSCAM were just in the commenting; I
noticed some glitches as I was wandering through the code.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@4996 10fe2890-8a40-4696-8b1d-7ccbb19c398e