Preliminary version of new MapPlot class that handles the new Ezmap dataset. Most dataset dependencies have been moved into a new private class hierarchy with superclass mapDataHandlerClass.
Access the new dataset by setting the resource mpDataBaseVersion to
the enumerated value "Ncarg4_1". You can switch back and forth after
creating a MapPlot. The new version supports the specified string resources;
however the strings are often not the same as the old version, so
switching back and forth between new and old may cause warning messages
if string specification resources are set.

For now you can only specify simple specified names for the new version.
You can retrieve the mpAreaNames resource to get a list of these. They
are handled in a case-insensitive manner. Unlike the old version, you
do not put a '-' character in between the words of multiword names.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5008 10fe2890-8a40-4696-8b1d-7ccbb19c398e